Aplikacja_Webowa_Etap_4/client/requirements.txt
flask
uuid
pyjwt
redis
requests
six
authlib
requests


